+++ experimental/packages/exporter-logs-otlp-grpc/src/OTLPLogExporter.ts
@@ -28,6 +28,11 @@ import {
   createExportLogsServiceRequest,
   IExportLogsServiceRequest,
 } from '@opentelemetry/otlp-transformer';
+import { VERSION } from './version';
+
+const USER_AGENT = {
+  'User-Agent': `OTel-OTLP-Exporter-JavaScript/${VERSION}`,
+};
 
 /**
  * OTLP Logs Exporter for Node
@@ -38,9 +43,12 @@ export class OTLPLogExporter
 {
   constructor(config: OTLPGRPCExporterConfigNode = {}) {
     super(config);
-    const headers = baggageUtils.parseKeyPairsIntoRecord(
-      getEnv().OTEL_EXPORTER_OTLP_LOGS_HEADERS
-    );
+    const headers = {
+      ...USER_AGENT,
+      ...baggageUtils.parseKeyPairsIntoRecord(
+        getEnv().OTEL_EXPORTER_OTLP_LOGS_HEADERS
+      ),
+    };
     this.metadata ||= new Metadata();
     for (const [k, v] of Object.entries(headers)) {
       this.metadata.set(k, v);